+++ app/Models/SealiacOverview.php
@@ -1,0 +1,26 @@
+<?php
+
+declare(strict_types=1);
+
+namespace App\Models;
+
+use App\Models\EatingOut\Eatery;
+use App\Models\EatingOut\NationwideBranch;
+use Illuminate\Database\Eloquent\Model;
+use Illuminate\Database\Eloquent\Relations\BelongsTo;
+use Illuminate\Database\Eloquent\Relations\MorphTo;
+
+class SealiacOverview extends Model
+{
+    public function model(): MorphTo
+    {
+        return $this->morphTo();
+    }
+
+    protected function casts(): array
+    {
+        return [
+            'invalidated' => 'boolean',
+        ];
+    }
+}
